Fix claim formula and disband not reflecting amounts properly
The issue is, if you've set a formula like "10000*c", you would expect
the c to be the current chunk you're claiming, so if it's the first one,
it should be c=1. With the old code it's c=0, so the first claim is
free.

The issue with disband is that it doesn't decrement the chunk count,
when paying back the player per chunk and when the fomula uses chunks,
the amount isn't the correct one.
Which means the amount the player receives is not the one they invested.
You create funds out of nothing with the "10000*c" formula this way,
it doesn't make sense.

A fix for the chunk count when sending the clans info to the player
is added to compensate for the new changes, because otherwise it will
show one chunk ahead.